Move visibility and mounting logic into camera stream controller
